% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output_format.R
\name{check_output_format}
\alias{check_output_format}
\title{Check the file extension of the output file to see if it is the same as the
`raster_format` or `vector_format` settings. If a raster, such as a GeoTIFF
is output directly from a SAGA-GIS tool but the raster format is set to SAGA,
then this might work depending on the saga version but Rsagacmd will not
know how to read the file.}
\usage{
check_output_format(x, raster_format, vector_format)
}
\arguments{
\item{x}{a `parameter` object that is an output parameter of a tool.}

\item{raster_format}{the raster format.}

\item{vector_format}{the vector format.}
}
\description{
Check the file extension of the output file to see if it is the same as the
`raster_format` or `vector_format` settings. If a raster, such as a GeoTIFF
is output directly from a SAGA-GIS tool but the raster format is set to SAGA,
then this might work depending on the saga version but Rsagacmd will not
know how to read the file.
}
\keyword{internal}
