% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output_format.R
\name{convert_sagaext_r}
\alias{convert_sagaext_r}
\title{Ensure that the file extension for the SAGA raster format ends with .sdat
for reading or writing SAGA grid objects in R.}
\usage{
convert_sagaext_r(fp)
}
\arguments{
\item{fp}{file path to raster writing}
}
\value{

}
\description{
This is used because the R raster/terra libraries expect to read and write
SAGA grid formats using the '.sdat' file extension, not '.sgrd'.
}
\keyword{internal}
