% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_alias}
\alias{create_alias}
\title{Generates a syntactically-correct R name based on a SAGA-GIS identifier}
\usage{
create_alias(identifier)
}
\arguments{
\item{identifier}{A character with the identifier.}
}
\value{
A character with a syntactically-correct alias.
}
\description{
SAGA-GIS identifiers sometimes cannot represent syntactically-correct names
in R because they start with numbers or have spaces. They are also all in
uppercase which is ugly to refer to in code. This function creates an
alternative/alias identifier.
}
\keyword{internal}
