% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-overrides.R
\name{create_tool_overrides}
\alias{create_tool_overrides}
\title{Apply manually-defined changes to specific tools}
\usage{
create_tool_overrides(tool_name, params)
}
\arguments{
\item{tool_name}{character, name of the tool. This is the alias name used by
Rsagacmd, i.e. the tool name without spaces, all lowercase etc.}

\item{params}{the `parameters` object for the tool}
}
\value{
the altered `parameters` object
}
\description{
Used to manually alter or add parameters for specific tools outside of what
has been defined based on the output of saga_cmd --create-docs
}
\keyword{internal}
