% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elevation}
\alias{elevation}
\title{Example elevation data}
\format{A RasterLayer object with 75 rows and 70 columns. For more details
refer to the GRASS GIS North Carolina sample data:
\describe{
  \item{elevation}{Altitude in masl.}
  ...
}}
\source{
\url{https://grass.osgeo.org/download/sample-data}
}
\usage{
elevation
}
\description{
A small tile of elevation data derived from the open-source
GRASS-GIS 7.x example North Carolina dataset
}
\keyword{datasets}
