% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{extract_tool}
\alias{extract_tool}
\title{Internal function to extract information from a `saga_tool` object}
\usage{
extract_tool(x)
}
\arguments{
\item{x}{a `saga_tool` object}
}
\value{
the intervals of a `saga_tool`
}
\description{
Internal function to extract information from a `saga_tool` object
}
\keyword{internal}
