% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsagacmd.R
\name{initSAGA}
\alias{initSAGA}
\title{Dynamically maps all SAGA-GIS functions to a nested list of functions in R.}
\usage{
initSAGA(saga_bin = NA)
}
\arguments{
\item{saga_bin}{Optional path to saga_cmd}
}
\value{
Nested list of functions for SAGA-GIS libraries and tools
#' @export
}
\description{
A simple wrapper that dynamically creates R functions to all valid tools and
returns them as a nested and named list of SAGA-GIS libraries and tools.
}
