% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MRVBFthreshold}
\alias{MRVBFthreshold}
\title{Calculate the T_SLOPE value based on DEM resolution for MRVBF}
\usage{
MRVBFthreshold(res, plot = FALSE)
}
\arguments{
\item{res}{Input DEM resolution}

\item{plot}{logical, produce plot of relationship}
}
\value{
Numeric. T_SLOPE value for MRVBF
}
\description{
Calculates the T_SLOPE value for the Multiresolution Index of Valley Bottom 
Flatness (Gallant and Dowling, 2003) based on input DEM resolution. MRVBF
identified valley bottoms based on classifying slope angle and identifying
low areas by ranking elevation in respect to the surrounding topography
across a range of DEM resolutions. The MRVBF algorithm was developed using a
25 m DEM, and so if the input DEM has a different resolution then the slope
threshold T_SLOPE needs to be adjusted from its default value of 16 in order
to maintain the relationship between slope and DEM resolution. This function
provides a convenient way to perform that calculation.
}
\examples{
\dontrun{
MRVBFthreshold(res = 10, plot = TRUE)
}
}
