% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{Parameter}
\alias{Parameter}
\title{Parameter class}
\description{
Stores metadata associated with each SAGA-GIS tool parameter.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{A character to describe the data type of the parameter such 
as "input", "output", "Grid", "Grid list", "Shapes", etc.}

\item{\code{name}}{Character with the long name of the parameter.}

\item{\code{alias}}{Syntactically correct alias for the identifier.}

\item{\code{identifier}}{Character with the identifier used by saga_cmd.}

\item{\code{description}}{A character with the description of the parameter.}

\item{\code{constraints}}{A character describing the parameters constraints.}

\item{\code{io}}{Either purpose of the parameters, either 'input' or 'output'.}

\item{\code{feature}}{The type of data that the parameter represents. Examples
include 'Grid', 'Grid list', 'Shapes' etc.}

\item{\code{default}}{The default value for the parameters.}

\item{\code{minumum}}{The minimum value for the parameter.}

\item{\code{maximum}}{The maximum value for the parameter.}

\item{\code{value}}{The assigned value for the parameter. This value will be
updated when the tool is run by the user.}

\item{\code{files}}{Any files (potentially after being written to disk) that
are associated with the parameter, for example an input grid.}

\item{\code{temp_path}}{Character specifying the tempdir to use for storage}

\item{\code{raster_format}}{Name of the raster format to use for saving data.
Must match a name in `supported_raster_formats`.}

\item{\code{vector_format}}{Name of the vector format to use for saving data.
Must match a name in `supported_vector_formats`.}

\item{\code{grid_features}}{Character vector of output types that represent grids}

\item{\code{shape_features}}{Character vector of output types that represent shapes}

\item{\code{table_features}}{Character vector of output types that represent tables}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{minumum}}{The minimum value for the parameter.}

\item{\code{grid_features}}{Character vector of output types that represent grids}

\item{\code{shape_features}}{Character vector of output types that represent shapes}

\item{\code{table_features}}{Character vector of output types that represent tables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-parameter-new}{\code{Parameter$new()}}
\item \href{#method-parameter-update_parameter_file}{\code{Parameter$update_parameter_file()}}
\item \href{#method-parameter-assign_tempfile}{\code{Parameter$assign_tempfile()}}
\item \href{#method-parameter-clone}{\code{Parameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameter-new"></a>}}
\if{latex}{\out{\hypertarget{method-parameter-new}{}}}
\subsection{Method \code{new()}}{
Create a new parameter object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$new(type, name, alias, identifier, description, constraints)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A character to describe the data type of the parameter.}

\item{\code{name}}{A character with the long name of the parameter.}

\item{\code{alias}}{A syntactically correct alias for the identifier.}

\item{\code{identifier}}{A character with the identifier of the parameter used
by saga_cmd.}

\item{\code{description}}{A character with the description of the parameter.}

\item{\code{constraints}}{A character describing the parameters constraints.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameter-update_parameter_file"></a>}}
\if{latex}{\out{\hypertarget{method-parameter-update_parameter_file}{}}}
\subsection{Method \code{update_parameter_file()}}{
Updates the `Parameter` object with file paths to the R data objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$update_parameter_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameter-assign_tempfile"></a>}}
\if{latex}{\out{\hypertarget{method-parameter-assign_tempfile}{}}}
\subsection{Method \code{assign_tempfile()}}{
Update a parameters file field with a path to a temporary file to
collect outputs from SAGA-GIS that are automatically created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$assign_tempfile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-parameter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
