% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Generates a list of `parameter` objects for a SAGA-GIS tool}
\usage{
parameters(tool_options)
}
\arguments{
\item{tool_options}{A data.frame containing the table that refers to the
SAGA-GIS tool parameter options.}
}
\value{
A `parameters` object
}
\description{
Each `parameter` object contains information about the datatype, permissible values
and input/output settings associated with each identifier for a SAGA-GIS tool.
}
\keyword{internal}
