% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.saga_tool}
\alias{print.saga_tool}
\title{Generic function to display help and usage information for any SAGA-GIS tool}
\usage{
\method{print}{saga_tool}(x, ...)
}
\arguments{
\item{x}{saga_tool object}

\item{...}{Additional arguments to pass to print. Currently not used.}
}
\description{
Generic function to display help and usage information for any SAGA-GIS tool
}
\examples{
\dontrun{
# Intialize a saga object
saga <- saga_gis()

# Display usage information on a tool
print(saga$ta_morphometry$Slope_Aspect_Curvature)

# Or simply:
saga$ta_morphometry$Slope_Aspect_Curvature
}
}
