% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{r_to_saga}
\alias{r_to_saga}
\title{Saves R objects to temporary files for processing by SAGA-GIS.}
\usage{
r_to_saga(param, temp_path)
}
\arguments{
\item{param}{A single variable that may be a raster object, sp object, sf object or dataframe}

\item{temp_path}{character. Path to directory to store temporary files.}
}
\value{
character. Character string of filename of saved object
}
\description{
Intended to be used internally. Raster objects are checked to see if the object is linked to a
file or exists only in memory. Spatial, sf objects and dataframes are saved to temporary files.
}
