% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_grid}
\alias{read_grid}
\title{Read a raster data set that is output by saga_cmd}
\usage{
read_grid(x, backend)
}
\arguments{
\item{x}{list, a `options` object that was created by the `create_tool`
function that contains the parameters for a particular tool and its
outputs.}

\item{backend}{character, either "raster" or "terra"}
}
\value{
either a `raster` or `SpatRaster` object
}
\description{
Read a raster data set that is output by saga_cmd
}
\keyword{internal}
