% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_shapes}
\alias{read_shapes}
\title{Read a spatial vector data set that is output by saga_cmd}
\usage{
read_shapes(x, vector_backend)
}
\arguments{
\item{x}{list, a `options` object that was created by the `create_tool`
function that contains the parameters for a particular tool and its
outputs.}

\item{vector_backend}{character for vector backend to use.}
}
\value{
an `sf` object.
}
\description{
Read a spatial vector data set that is output by saga_cmd
}
\keyword{internal}
