% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-read.R
\name{read_table}
\alias{read_table}
\title{Read a tabular data set that is output by saga_cmd}
\usage{
read_table(x)
}
\arguments{
\item{x}{list, a `options` object that was created by the `create_tool`
function that contains the parameters for a particular tool and its
outputs.}
}
\value{
a `tibble`.
}
\description{
Read a tabular data set that is output by saga_cmd
}
\keyword{internal}
