% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-cmd.R
\name{run_cmd}
\alias{run_cmd}
\title{Prepares the statement and runs the external saga_cmd executable}
\usage{
run_cmd(saga_cmd, saga_config, lib, tool_cmd, args, verbose)
}
\arguments{
\item{saga_cmd}{character, name of the saga_cmd executable or alias.}

\item{saga_config}{character, path to the saga configuration "ini" file.}

\item{lib}{character, name of the selected library.}

\item{tool_cmd}{character, name of the selected tool.}

\item{args}{named list of tool options, such as list(DEM = "mygrid.tif",
RADIUS = 3).}

\item{verbose}{logical, whether to show all saga_cmd messages on the R
console.}
}
\value{
list, output from `processx::run()`
}
\description{
Prepares the statement and runs the external saga_cmd executable
}
\keyword{internal}
