% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shortcuts.R
\name{saga.Fill_Sinks}
\alias{saga.Fill_Sinks}
\title{Sink Removal}
\usage{
saga.Fill_Sinks(DEM, SINKROUTE = NA, DEM_PREPROC = NA, METHOD = NA,
  THRESHOLD = NA, THRSHEIGHT = NA, senv, intern = TRUE, usage = FALSE)
}
\arguments{
\item{DEM}{DEM: Grid (input). Digital Elevation Model that has to be
processed}

\item{SINKROUTE}{Sink Route: Grid (optional input).}

\item{DEM_PREPROC}{Preprocessed DEM: Grid (optional output).}

\item{METHOD}{Method: Choice. Available Choices:
\itemize{
\item [0] Deepen Drainage Routes
\item [1] Fill Sinks
\item Default: 1
}}

\item{THRESHOLD}{Threshold: Boolean. Default: 0}

\item{THRSHEIGHT}{Threshold Height: Floating point. The parameter describes
the maximum depth of a sink to be considered for removal [map units]. This
makes it possible to exclude deeper sinks from filling.}

\item{senv}{SAGA-GIS environment returned by initSAGA}

\item{intern}{Optionally load output as an R object; Default is TRUE}

\item{usage}{Boolean, display tool help}
}
\value{
Specified SAGA-GIS outputs
}
\description{
Fill/remove sinks within a digital elevation model
}
