% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shortcuts.R
\name{saga.Flow_Accumulation}
\alias{saga.Flow_Accumulation}
\title{Flow Accumulation (Top-Down)}
\usage{
saga.Flow_Accumulation(ELEVATION, SINKROUTE = NA, WEIGHTS = NA,
  VAL_INPUT = NA, ACCU_MATERIAL = NA, ACCU_TARGET, LINEAR_VAL = NA,
  LINEAR_DIR = NA, FLOW, VAL_MEAN, ACCU_TOTAL = NA, ACCU_LEFT = NA,
  ACCU_RIGHT = NA, FLOW_LENGTH = NA, WEIGHT_LOSS = NA, STEP = NA,
  FLOW_UNIT = NA, METHOD = NA, LINEAR_DO = NA, LINEAR_MIN = NA,
  CONVERGENCE = NA, NO_NEGATIVES = NA, senv, intern = TRUE,
  usage = FALSE)
}
\arguments{
\item{ELEVATION}{Elevation: Grid (input)}

\item{SINKROUTE}{Sink Routes: Grid (optional input)}

\item{WEIGHTS}{Weights: Grid (optional input)}

\item{VAL_INPUT}{Input for Mean over Catchment: Grid (optional input)}

\item{ACCU_MATERIAL}{Material for Accumulation: Grid (optional input)}

\item{ACCU_TARGET}{Accumulation Target: Grid (input)}

\item{LINEAR_VAL}{Linear Flow Threshold Grid: Grid (optional input). Optional
grid providing values to be compared with linear flow threshold instead of
flow accumulation}

\item{LINEAR_DIR}{Channel Direction: Grid (optional input). Use this for
(linear) flow routing, if the value is a valid direction (0-7 = N, NE, E,
SE, S, SW, W, NW)}

\item{FLOW}{Flow Accumulation: Grid (output)}

\item{VAL_MEAN}{Mean over Catchment: Grid (output)}

\item{ACCU_TOTAL}{Accumulated Material: Grid (optional output)}

\item{ACCU_LEFT}{Accumulated Material (Left Side): Grid (optional output)}

\item{ACCU_RIGHT}{Accumulated Material (Right Side): Grid (optional output)}

\item{FLOW_LENGTH}{Flow Path Length: Grid (optional output). Average distance
that a cell's accumulated flow travelled}

\item{WEIGHT_LOSS}{Loss through Negative Weights: Grid (optional output).
When using weights without support for negative flow: output of the
absolute amount of negative flow that occurred}

\item{STEP}{Step: Integer. For testing purposes. Only generate flow at cells
with step distance (each step row/column). Minimum: 1; Default: 1}

\item{FLOW_UNIT}{Flow Accumulation Unit: Choice. Available Choices:
\itemize{
\item [0] number of cells
\item [1] cell area
\item Default: 1
}}

\item{METHOD}{Method: Choice. Available Choices:
\itemize{
\item [0] Deterministic 8
\item [1] Rho 8
\item [2] Braunschweiger Reliefmodell
\item [3] Deterministic Infinity
\item [4] Multiple Flow Direction
\item [5] Multiple Triangular Flow Directon
\item [6] Multiple Maximum Downslope Gradient Based Flow Directon
\item Default: 4
}}

\item{LINEAR_DO}{Thresholded Linear Flow: Boolean. Apply linear flow routing
(D8) to all cells, having a flow accumulation greater than the specified
threshold. Default: 1}

\item{LINEAR_MIN}{Linear Flow Threshold: Integer. Flow accumulation threshold
(cells) for linear flow routing. Minimum: 0. Default: 500}

\item{CONVERGENCE}{Convergence: Floating point. Convergence factor for
Multiple Flow Direction Algorithm (Freeman 1991). Applies also to the
Multiple Triangular Flow Directon Algorithm. Minimum: 0.000000. Default:
1.100000}

\item{NO_NEGATIVES}{Prevent Negative Flow Accumulation: Boolean. when using
weights: do not transport negative flow, set it to zero instead; useful
e.g. when accumulating measures of water balance. Default: 1}

\item{senv}{SAGA-GIS environment returned by initSAGA}

\item{intern}{Optionally load output as an R object; Default is TRUE}

\item{usage}{Boolean, display tool help}
}
\value{
Specified SAGA-GIS outputs
}
\description{
Top-down processing of cells for calculation of flow accumulation and related
parameters. This set of algorithms processes a DEM downwards from the highest
to the lowest cell. Flow routing methods provided by this tool:
\itemize{
\item Deterministic 8 (aka D8, O'Callaghan & Mark 1984)
\item Braunschweiger Reliefmodell (Bauer et al. 1985)
\item Rho 8 (Fairfield & Leymarie 1991)
\item Multiple Flow Direction (Freeman 1991, Quinn et al. 1991)
\item Deterministic Infinity (Tarboton 1997)
\item Triangular Multiple Flow Direction (Seibert & McGlynn 2007
\item Multiple Flow Direction based on Maximum Downslope Gradient (Qin et al. 2011)
}
}
