% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shortcuts.R
\name{saga.Mosaicking}
\alias{saga.Mosaicking}
\title{Mosaicking}
\usage{
saga.Mosaicking(GRIDS, TARGET_TEMPLATE = NA, TARGET_OUT_GRID, NAME = NA,
  TYPE = NA, RESAMPLING = NA, OVERLAP = NA, BLEND_DIST = NA,
  MATCH = NA, TARGET_DEFINITION = NA, TARGET_USER_SIZE = NA,
  TARGET_USER_XMIN = NA, TARGET_USER_XMAX = NA, TARGET_USER_YMIN = NA,
  TARGET_USER_YMAX = NA, TARGET_USER_FITS = NA, senv, intern = TRUE,
  usage = FALSE)
}
\arguments{
\item{GRIDS}{Grids: Grid list (input)}

\item{TARGET_TEMPLATE}{Target System: Grid (optional input)}

\item{TARGET_OUT_GRID}{Target Grid: Grid (output)}

\item{NAME}{Name: Text. Default: Mosaic}

\item{TYPE}{Data Storage Type: Choice. Available Choices:
\itemize{
\item [0] 1 bit
\item [1] 1 byte unsigned integer
\item [2] 1 byte signed integer
\item [3] 2 byte unsigned integer
\item [4] 2 byte signed integer
\item [5] 4 byte unsigned integer
\item [6] 4 byte signed integer
\item [7] 4 byte floating point
\item [8] 8 byte floating point
\item [9] same as first grid in list
\item Default: 9
}}

\item{RESAMPLING}{Resampling: Choice. Available Choices:
\itemize{
\item [0] Nearest Neighbour
\item [1] Bilinear Interpolation
\item [2] Bicubic Spline Interpolation
\item [3] B-Spline Interpolation
\item Default: 3
}}

\item{OVERLAP}{Overlapping Areas: Choice. Available Choices:
\itemize{
\item [0] first
\item [1] last
\item [2] minimum
\item [3] maximum
\item [4] mean
\item [5] blend boundary
\item [6] feathering
}}

\item{BLEND_DIST}{Blending Distance: Floating point. Minimum: 0.000000; Default: 10.000000}

\item{MATCH}{Match: Choice. Available Choices:
\itemize{
\item [0] none
\item [1] match histogram of first grid in list
\item [2] match histogram of overlapping area
\item [3] regression
\item Default: 0
}}

\item{TARGET_DEFINITION}{Target Grid System: Choice. Available Choices:
\itemize{
\item [0] user defined
\item [1] grid or grid system
\item Default: 0
}}

\item{TARGET_USER_SIZE}{Cellsize: Floating point. Minimum: 0.000000; Default: 1.000000}

\item{TARGET_USER_XMIN}{Left: Floating point. Default: 0.000000}

\item{TARGET_USER_XMAX}{Right: Floating point. Default: 100.000000}

\item{TARGET_USER_YMIN}{Bottom: Floating point. Default: 0.000000}

\item{TARGET_USER_YMAX}{Top: Floating point. Default: 100.000000}

\item{senv}{SAGA-GIS environment returned by initSAGA}

\item{intern}{Optionally load output as an R object; Default is TRUE}

\item{usage}{Boolean, display tool help}
}
\value{
Specified SAGA-GIS outputs
}
\description{
Merges multiple grids into one single grid.
}
