% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsagacmd.R
\name{saga.showTmpFiles}
\alias{saga.showTmpFiles}
\title{List temporary files created by Rsagacmd}
\usage{
saga.showTmpFiles()
}
\value{
returns the file names of the files in the temp directory that have
been generated by Rsagacmd. Note this list of files only includes the primary
file extension, i.e. '.shp' for a shapefile without the accessory files
(e.g. .prj, .shx etc.).
}
\description{
For convenience, functions in the Rsagacmd package create temporary files if
any required outputs for a SAGA-GIS tool are not specified as arguments.
Temporary files in R are automatically removed at the end of each session.
However, when dealing with raster data, these temporary files potentially can
consume large amounts of disk space. These temporary files can be observed
during a session by using the saga.showTmpFiles function, and can be removed
using the saga.removeTmpFiles function.
}
