% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shortcuts.R
\name{saga.Slope_Aspect_Curvature}
\alias{saga.Slope_Aspect_Curvature}
\title{Slope, Aspect, Curvature}
\usage{
saga.Slope_Aspect_Curvature(ELEVATION, SLOPE, ASPECT, C_GENE = NA,
  C_PROF = NA, C_PLAN = NA, C_TANG = NA, C_LONG = NA, C_CROS = NA,
  C_MINI = NA, C_MAXI = NA, C_TOTA = NA, C_ROTO = NA, METHOD = NA,
  UNIT_SLOPE = NA, UNIT_ASPECT = NA, senv, intern = TRUE, usage = FALSE)
}
\arguments{
\item{ELEVATION}{Elevation: Grid (input)}

\item{SLOPE}{Slope: Grid (output)}

\item{ASPECT}{Aspect: Grid (output)}

\item{C_GENE}{General Curvature: Grid (optional output)}

\item{C_PROF}{Profile Curvature: Grid (optional output)}

\item{C_PLAN}{Plan Curvature: Grid (optional output)}

\item{C_TANG}{Tangential Curvature: Grid (optional output)}

\item{C_LONG}{Longitudinal Curvature: Grid (optional output)}

\item{C_CROS}{Cross-Sectional Curvature: Grid (optional output)}

\item{C_MINI}{Minimal Curvature: Grid (optional output)}

\item{C_MAXI}{Maximal Curvature: Grid (optional output)}

\item{C_TOTA}{Total Curvature: Grid (optional output)}

\item{C_ROTO}{Flow Line Curvature: Grid (optional output)}

\item{METHOD}{Method: Choice. Available choices:
\itemize{
\item [0] maximum slope (Travis et al. 1975)
\item [1] maximum triangle slope (Tarboton 1997)
\item [2] least squares fitted plane (Horn 1981, Costa-Cabral & Burgess 1996)
\item [3] 6 parameter 2nd order polynom (Evans 1979)
\item [4] 6 parameter 2nd order polynom (Heerdegen & Beran 1982)
\item [5] 6 parameter 2nd order polynom (Bauer, Rohdenburg, Bork 1985)
\item [6] 9 parameter 2nd order polynom (Zevenbergen & Thorne 1987)
\item [7] 10 parameter 3rd order polynom (Haralick 1983)
\item Default: 6
}}

\item{UNIT_SLOPE}{Slope units: Choice. Available choices:
\itemize{
\item [0] radians
\item [1] degree
\item [2] percent
\item Default: 0
}}

\item{UNIT_ASPECT}{Aspect Units: Choice. Available choices:
\itemize{
\item [0] radians
\item [1] degree
\item Default: 0
}}

\item{senv}{SAGA-GIS environment returned by initSAGA}

\item{intern}{Optionally load output as an R object; Default is TRUE}

\item{usage}{Boolean, display tool help}
}
\value{
Specified SAGA-GIS outputs
}
\description{
Calculates the local morphometric terrain parameters slope, aspect and if
supported by the chosen method also the curvature. Besides tangential
curvature also its horizontal and vertical components (i.e. plan and profile
curvature) can be calculated.
}
