% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shortcuts.R
\name{saga.SplitLayers}
\alias{saga.SplitLayers}
\title{Split a RasterStack or StackBrick into multiple bands}
\usage{
saga.SplitLayers(x)
}
\arguments{
\item{x}{RasterStack or RasterBrick}
}
\value{
list of RasterLayers
}
\description{
Split a RasterStack or StackBrick into multiple bands
}
