% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{saga_docs}
\alias{saga_docs}
\title{Browse the online documentation for a saga_tool}
\usage{
saga_docs(saga_tool)
}
\arguments{
\item{saga_tool}{a saga_tool object}
}
\description{
Browse the online documentation for a saga_tool
}
\examples{
\dontrun{
library(Rsagacmd)

saga <- saga_gis()

saga_docs(saga$ta_morphometry$slope_aspect_curvature)
}
}
