% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_execute.R
\name{saga_execute}
\alias{saga_execute}
\title{Function to execute SAGA-GIS commands through the command line tool}
\usage{
saga_execute(lib, tool, senv, .intern = TRUE, .all_outputs = TRUE, ...)
}
\arguments{
\item{lib}{character, name of SAGA-GIS library to execute}

\item{tool}{character, name of SAGA-GIS tool to execute}

\item{senv}{saga environment object object}

\item{.intern}{logical, whether to load the outputs from the SAGA-GIS
geoprocessing operatioon as an R object}

\item{.all_outputs}{logical, default = TRUE, automatically output all
results from the selected SAGA tool and load them results as R objects. If
.all_outputs = FALSE then the file paths to store the tool's results will
have to be manually specified in the arguments}

\item{...}{named arguments and values for SAGA tool}
}
\value{
output of SAGA-GIS tool loaded as an R object 
(RasterLayer/sf/dataframe)
}
\description{
Intended to be used internally
}
