% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{saga_search}
\alias{saga_search}
\title{Automatic search for the path to a SAGA-GIS installation}
\usage{
saga_search()
}
\value{
character. Path to installed saga_cmd binary
}
\description{
Returns the path to the saga_cmd executable On windows, automatic searching
will occur first in 'C:/Program Files/SAGA-GIS'; 'C:/Program Files
(x86)/SAGA-GIS'; 'C:/SAGA-GIS'; 'C:/OSGeo4W'; and 'C:/OSGeo4W64'. On linux or
OS X, saga_cmd is usually included in PATH, if not an automatic search is
performed in the '/usr' folder. If multiple versions of SAGA-GIS are
installed on the system, the path to the newest version is returned.
}
