% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{saga_show_tmpfiles}
\alias{saga_show_tmpfiles}
\title{List temporary files created by Rsagacmd}
\usage{
saga_show_tmpfiles()
}
\value{
returns the file names of the files in the temp directory that have
been generated by Rsagacmd. Note this list of files only includes the primary
file extension, i.e. '.shp' for a shapefile without the accessory files
(e.g. .prj, .shx etc.).
}
\description{
For convenience, functions in the Rsagacmd package create temporary files if
any outputs for a SAGA-GIS tool are not specified as arguments.
Temporary files in R are automatically removed at the end of each session.
However, when dealing with raster data, these temporary files potentially can
consume large amounts of disk space. These temporary files can be observed
during a session by using the saga_show_tmpfiles function, and can be removed
using the saga_remove_tmpfiles function.
}
\examples{
\dontrun{
# Show all temporary files generated by Rsagacmd
saga_remove_tmpfiles(h = 0)
}
}
