% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_version.R
\name{saga_version}
\alias{saga_version}
\title{Return the installed version of SAGA-GIS}
\usage{
saga_version(saga_cmd)
}
\arguments{
\item{saga_cmd}{The path of the saga_cmd binary.}
}
\value{
A numeric_version wit hthe version of SAGA-GIS found at the cmd path.
}
\description{
Intended to be used internally by \code{\link{saga_env}}. Uses a system call
to saga_cmd to output version of installed SAGA-GIS on the console
}
