% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaConfigure}
\alias{sagaConfigure}
\title{Generates a custom saga_cmd configuration file}
\usage{
sagaConfigure(senv, grid_caching = FALSE, grid_cache_threshlod = 100,
  grid_cache_dir = NA, cores = NA, sagaversion)
}
\arguments{
\item{senv}{sagaInstallation object. SAGA-GIS environment and settings}

\item{grid_caching}{logical. Optionally use file caching}

\item{grid_cache_threshlod}{numeric. Threshold (in Mb) before file caching
for loaded raster data is activated}

\item{grid_cache_dir}{character. Path to directory for temporary files}

\item{cores}{numeric. Maximum number of processing cores. Needs to be set to
1 if file caching is activated}

\item{sagaversion}{numeric_version. Version of SAGA-GIS. The generation of a
saga_cmd configuration file is only valid for versions > 4.0.0}
}
\value{
character. Path to custom saga_cmd initiation file
}
\description{
Creates and edits a saga_cmd coniguration file in order to change saga_cmd
settings related to file caching and number of available processor cores.
Intended to be used internally by \code{\link{sagaGIS}}.
}
