% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsagacmd.R
\name{sagaEnv}
\alias{sagaEnv}
\title{Establishes the link to SAGA GIS by generating a SAGA help file and parsing
all libraries, tools and options from the help files into a nested list of 
library, module and options}
\usage{
sagaEnv(saga_bin = NA)
}
\arguments{
\item{saga_bin}{Optional character with known path to saga_cmd binary}
}
\value{
List of saga_cmd path, SAGA-GIS version and nested list of libaries tools and options
}
\description{
Establishes the link to SAGA GIS by generating a SAGA help file and parsing
all libraries, tools and options from the help files into a nested list of 
library, module and options
}
