% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaExecute}
\alias{sagaExecute}
\title{Function to execute SAGA-GIS commands through the command line tool.}
\usage{
sagaExecute(lib, tool, senv, intern = TRUE, ...)
}
\arguments{
\item{lib}{character. Name of SAGA-GIS library to execute}

\item{tool}{character. Name of SAGA-GIS tool to execute}

\item{senv}{sagaInstallation object}

\item{intern}{logical. Load outputs as R objects}

\item{...}{Named arguments and values for SAGA tool}
}
\value{
Output of SAGA-GIS tool loaded as an R object
  (RasterLayer/sf/dataframe)
}
\description{
Intended to be used internally
}
