% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaExtract}
\alias{sagaExtract}
\title{Extract raster data at vector feature locations using SAGA-GIS}
\usage{
sagaExtract(x, SHAPES, GRIDS, RESAMPLING = 0)
}
\arguments{
\item{x}{sagaGIS object}

\item{SHAPES}{sf or sp vector object}

\item{GRIDS}{list of RasterLayer, RasterBrick, or RasterStack objects
containing only a single layer}

\item{RESAMPLING}{numeric or character. Resampling method used to extract
pixel values. [0] Nearest Neighbour;[1] Bilinear Interpolation;[2] Bicubic
Spline Interpolation;[3] B-Spline Interpolation}
}
\value{
sf object containing extracted raster data
}
\description{
Extracts pixel values at vector feature locations using SAGA-GIS. If the
vector feature represents point or lines, then a single pixel value is
extracted. For polygon data, the average pixel value of each grid within the
area of the polygon is extracted.
}
