% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsagacmd.R
\name{sagaGeo}
\alias{sagaGeo}
\title{Main function to execute SAGA-GIS commands through the command line tool}
\usage{
sagaGeo(lib, tool, senv, intern = TRUE, ...)
}
\arguments{
\item{lib}{Character string of name of SAGA-GIS library to execute}

\item{tool}{Character string of name of SAGA-GIS tool to execute}

\item{senv}{SAGA-GIS environment setting}

\item{intern}{Boolean to load outputs as R objects}

\item{...}{Named arguments and values for SAGA tool}
}
\value{
Output of SAGA-GIS tool loaded as an R object (raster/rasterstack/sf/dataframe)
}
\description{
Main function to execute SAGA-GIS commands through the command line tool
}
