% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaSearch}
\alias{sagaSearch}
\title{Automatic search for the path to a SAGA-GIS installation}
\usage{
sagaSearch()
}
\value{
character. Path to installed saga_cmd binary
}
\description{
Returns the path to the saga_cmd executable On windows, automatic searching
will search in first in C:/Program Files/SAGA-GIS; C:/Program Files
(x86)/SAGA-GIS; C:/SAGA-GIS; 'C:/OSGeo4W', and 'C:/OSGeo4W64'. On linux or OS
X, saga_cmd is usually included in PATH, if not an automatic search is
performed in the /usr folder. If multiple versions of SAGA-GIS are installed
on the system, the path to the newest version is returned. Intended to be
called internally by the SAGA function.
}
