% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaVersion}
\alias{sagaVersion}
\title{Return the installed version of SAGA-GIS.}
\usage{
sagaVersion(saga_cmd)
}
\arguments{
\item{saga_cmd}{character. Path of the saga_cmd binary}
}
\value{
numeric_version. Version of SAGA-GIS found at the cmd path
}
\description{
Intended to be used internally by \code{\link{sagaEnv}}
}
