% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{save_object}
\alias{save_object}
\title{Generic methods to save R in-memory objects to file to SAGA-GIS to access}
\usage{
save_object(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{Other parameters such as the temporary directory to use for
storage.}
}
\value{
A character that specifies the file path to where the R object was
  saved.
}
\description{
Designed to be used internally by Rsagacmd for automatically pass data to
SAGA-GIS for geoprocessing
}
