% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_tools}
\alias{search_tools}
\title{Search for a SAGA-GIS tool}
\usage{
search_tools(x, pattern)
}
\arguments{
\item{x}{saga object}

\item{pattern}{character, pattern of text to search for within the tool name}
}
\value{
dataframe, tools that match the pattern of the search text and
their host library
}
\description{
Search for a SAGA-GIS tool
}
\examples{
\dontrun{
# initialize Rsagacmd
saga <- saga_gis()

# search for a tool
search_tools(x = saga, pattern = "terrain")
}
}
