% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsagacmd.R
\name{searchSAGA}
\alias{searchSAGA}
\title{Function to attempt to automatically find and select newest SAGA-GIS installation}
\usage{
searchSAGA()
}
\value{
Full path to installed saga_cmd binary
}
\description{
Function to attempt to automatically find and select newest SAGA-GIS installation
}
