% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{searchTools}
\alias{searchTools}
\title{Search for a SAGA-GIS tool}
\usage{
searchTools(x, pattern)
}
\arguments{
\item{x}{saga object}

\item{pattern}{character. Pattern of text to search for within the tool name}
}
\value{
dataframe of tools that match the pattern of the search text and
their host library
}
\description{
Search for a SAGA-GIS tool
}
\examples{
# initialize Rsagacmd
saga = sagaGIS()

# search for a tool
searchTools(x = saga, pattern = 'Terrain')
}
