% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_tools.R
\name{summarize_tool_params}
\alias{summarize_tool_params}
\title{Interval function used to summarize a `saga_tool` into a tibble that
describes the tools parameters and options}
\usage{
summarize_tool_params(tool_obj)
}
\arguments{
\item{tool_obj}{a nested list which constitutes the internals of a saga_tool
object}
}
\value{
a tibble
}
\description{
Interval function used to summarize a `saga_tool` into a tibble that
describes the tools parameters and options
}
\keyword{internal}
