% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.saga}
\alias{tidy.saga}
\title{Summarize the libraries that are available within a saga object and
return these as a tibble.}
\usage{
\method{tidy}{saga}(x, ...)
}
\arguments{
\item{x}{a `saga` object}

\item{...}{additional arguments. Currently unused.}
}
\value{
a tibble that describes libraries, their descriptions and number of
  tools that are available in SAGA-GIS.
}
\description{
Summarize the libraries that are available within a saga object and
return these as a tibble.
}
\examples{
\dontrun{
# Initialize a saga object
saga <- saga_gis()

# tidy the tools parameters into a tibble
tidy(saga)
}
}
