% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.saga_library}
\alias{tidy.saga_library}
\title{Summarize the tools that are available within a saga library and
return these as a tibble.}
\usage{
\method{tidy}{saga_library}(x, ...)
}
\arguments{
\item{x}{a `saga_library` object}

\item{...}{additional arguments. Currently unused.}
}
\value{
a tibble that describes the tools and their descriptions within a
  particular SAGA-GIS library.
}
\description{
Summarize the tools that are available within a saga library and
return these as a tibble.
}
\examples{
\dontrun{
# Initialize a saga object
saga <- saga_gis()

# tidy the library's parameters into a tibble
tidy(saga$climate_tools)
}
}
