% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{update_parameters_file}
\alias{update_parameters_file}
\title{Updates a `parameters` object with file paths to the R data objects.}
\usage{
update_parameters_file(params, temp_path = NULL, backend = NULL)
}
\arguments{
\item{temp_path}{A character specifying the tempdir to use for storage (optional).}

\item{backend}{A character with the raster backend ('raster' or 'terra).}

\item{param}{A `parameters` object.}

\item{all_outputs}{A logical indicating whether to use tempfiles for unspecified outputs.}
}
\value{
A `parameters` object with updated `file` attributes that refers to
  the on-disk file for saga_cmd to access.
}
\description{
Updates a `parameters` object with file paths to the R data objects.
}
\keyword{internal}
