% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tool.R
\name{create_tool}
\alias{create_tool}
\title{Generates list of options for a SAGA-GIS tool}
\usage{
create_tool(tool_information, tool_options)
}
\arguments{
\item{tool_information}{list}

\item{tool_options}{list}
}
\value{
A `saga_tool` object containing:
+ `tool_name` A syntactically-correct name for the tool.
+ `tool_cmd` The command to use for saga_cmd to execute tool.
+ `parameters` A named list of the tool's parameter objects.
}
\description{
Parses the html table for a SAGA-GIS tool into a list of identifiers,
options, defaults and constraints
}
\keyword{internal}
