% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initSAGA}
\alias{initSAGA}
\title{Dynamically maps all SAGA-GIS functions to a nested list of functions in R.}
\usage{
initSAGA(saga_bin = NA)
}
\arguments{
\item{saga_bin}{Optional path to saga_cmd}
}
\value{
Nested list of functions for SAGA-GIS libraries and tools
}
\description{
Dynamically creates R functions to all valid tools and
returns them as a nested and named list of SAGA-GIS libraries and tools. 
This list is defined as an S3 SAGAgis object, and the enclosed tools
represent SAGAtool objects.
}
\examples{
\dontrun{
# initialize Rsagacmd and dynamically generate functions for all SAGA-GIS tools
saga = initSAGA()

# Example of terrain analysis

# Generate random terrain and load as raster object
dem = saga$grid_calculus$Random_Terrain(RADIUS=100)

# Display help on usage for tool
print(saga$ta_morphometry$Terrain_Ruggedness_Index_TRI)

# Use Rsagacmd for to calculate the terrain ruggedness index
tri = saga$ta_morphometry$Terrain_Ruggedness_Index_TRI(DEM = dem)
plot(tri)

# Do not load output as an R object
saga$ta_morphometry$Terrain_Ruggedness_Index_TRI(DEM = dem, intern=FALSE)
}
}
