% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsagacmd.R
\name{initSAGA}
\alias{initSAGA}
\title{Dynamically maps all SAGA-GIS functions to a nested list of functions in R.
Returns a nested list containing function of all SAGA-GIS libraries and tools}
\usage{
initSAGA(saga_bin = NA)
}
\arguments{
\item{saga_bin}{optional path to saga_cmd}
}
\value{
nested list of functions for SAGA-GIS libraries and tools
}
\description{
Dynamically maps all SAGA-GIS functions to a nested list of functions in R.
Returns a nested list containing function of all SAGA-GIS libraries and tools
}
