% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.sagaTool}
\alias{print.sagaTool}
\title{Generic function to display help and usage information for any SAGA-GIS tool}
\usage{
print.sagaTool(x, ...)
}
\arguments{
\item{x}{sagaTool object}

\item{...}{Additional arguments to pass to print}
}
\description{
Generic function to display help and usage information for any SAGA-GIS tool
}
\examples{
# Initialize Rsagacmd if SAGA-GIS is installed in a recognized location
if(!is.null(sagaSearch())) {
    saga = sagaGIS()
    
    # Display usage information on a tool
    print(saga$ta_morphometry$Slope_Aspect_Curvature)
    
    # Or simply
    saga$ta_morphometry$Slope_Aspect_Curvature
}
}
