% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{saga_env}
\alias{saga_env}
\title{Parses valid SAGA-GIS libraries and tools into a nested list of functions}
\usage{
saga_env(saga_bin = NULL, opt_lib = NULL)
}
\arguments{
\item{saga_bin}{character. Optional path to saga_cmd executable}

\item{opt_lib}{character vector. List of subset of SAGA-GIS tool libraries
to generate links to.}
}
\value{
saga environment object object. Contains paths, settings and a nested list of libaries
  tools and options
}
\description{
Establishes the link to SAGA GIS by generating a SAGA help file and parsing all libraries, tools
and options from the help files into a nested list of library, module and options, that are
contained within an saga environment object object. Intended to be used internally by
\code{\link{saga_gis}}
}
