% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{saga_execute}
\alias{saga_execute}
\title{Function to execute SAGA-GIS commands through the command line tool.}
\usage{
saga_execute(lib, tool, senv, intern = TRUE, all_outputs = TRUE, ...)
}
\arguments{
\item{lib}{character. Name of SAGA-GIS library to execute}

\item{tool}{character. Name of SAGA-GIS tool to execute}

\item{senv}{saga environment object object}

\item{intern}{logical. Load outputs as R objects}

\item{all_outputs}{logical (default = TRUE). Automatically output all results
from the selected SAGA tool and load them results as R objects. If
all_outputs = FALSE then the file paths to store the tool's results will
have to be manually specified in the arguments}

\item{...}{Named arguments and values for SAGA tool}
}
\value{
Output of SAGA-GIS tool loaded as an R object
  (RasterLayer/sf/dataframe)
}
\description{
Intended to be used internally
}
