% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaEnv}
\alias{sagaEnv}
\title{Parses valid SAGA-GIS libraries and tools into a nested list of functions.}
\usage{
sagaEnv(saga_bin = NA)
}
\arguments{
\item{saga_bin}{Optional character with known path to saga_cmd binary}
}
\value{
List of saga_cmd path, SAGA-GIS version and nested list of libaries
  tools and options
}
\description{
Establishes the link to SAGA GIS by generating a SAGA help file and parsing
all libraries, tools and options from the help files into a saga.gis class
object, which contains a nested list of library, module and options.
Intended to be used internally and called by the initSAGA function.
}
