% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaGeo}
\alias{sagaGeo}
\title{Main function to execute SAGA-GIS commands through the command line tool.}
\usage{
sagaGeo(lib, tool, senv, intern = TRUE, cores = NULL, ...)
}
\arguments{
\item{lib}{Character string of name of SAGA-GIS library to execute}

\item{tool}{Character string of name of SAGA-GIS tool to execute}

\item{senv}{SAGA-GIS environment setting}

\item{intern}{Boolean to load outputs as R objects}

\item{cores}{Number of physical processing cores to use for computation. Default uses all available cores}

\item{...}{Named arguments and values for SAGA tool}
}
\value{
Output of SAGA-GIS tool loaded as an R object (RasterLayer/sf/dataframe)
}
\description{
Intended to be used internally, although sagaGeo can be used directly by
passing the name of the SAGA-GIS library and tool, along with named
arguments and values.
}
\examples{
\dontrun{
# initialize Rsagacmd and dynamically generate functions for all SAGA-GIS tools
saga = initSAGA()

# Example of running a SAGA-GIS tool

# Generate random terrain and load as raster object
dem = saga$grid_calculus$Random_Terrain()
dem = sagaGeo(lib = 'grid_calculus', tool = 'Random Terrain', senv = saga, 
RADIUS = 3, TARGET_OUT_GRID = 'somegrid.sgrd')
}
}
