% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaGeo}
\alias{sagaGeo}
\title{Main function to execute SAGA-GIS commands through the command line tool.}
\usage{
sagaGeo(lib, tool, saga_cmd, tool_cmd, tool_params, intern = TRUE,
  cores = NULL, ...)
}
\arguments{
\item{lib}{character. Name of SAGA-GIS library to execute}

\item{tool}{Character. Name of SAGA-GIS tool to execute}

\item{saga_cmd}{character. Path for the saga_cmd executable}

\item{tool_cmd}{character. SAGA-GIS command required to execute the tool}

\item{tool_params}{data.frame. DataFrame containing arguments and inputs/
outputs of the SAGA-GIS tool.}

\item{intern}{logical. Load outputs as R objects}

\item{cores}{numeric. Number of physical processing cores to use for
computation. Default uses all available cores}

\item{...}{Named arguments and values for SAGA tool}
}
\value{
Output of SAGA-GIS tool loaded as an R object
  (RasterLayer/sf/dataframe)
}
\description{
Intended to be used internally, although sagaGeo can be used directly by
passing the name of the SAGA-GIS library and tool, along with named
arguments and values.
}
