% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sagaGIS}
\alias{sagaGIS}
\title{Initiate a SAGA-GIS geoprocessor object}
\usage{
sagaGIS(saga_bin = NA, grid_caching = FALSE, grid_cache_threshlod = 100,
  grid_cache_dir = NA, cores = NA)
}
\arguments{
\item{saga_bin}{character, optional. Path to saga_cmd executable. If this 
argument is not supplied then an automatic search for the saga_cmd executable
will be performed.}

\item{grid_caching}{logical, optional. Use file caching in saga_cmd
geoprocessing operations for rasters that are too large to fit into memory.}

\item{grid_cache_threshlod}{numeric, optional. Threshold (in Mb) before file
caching is activated for loaded raster data.}

\item{grid_cache_dir}{character, optional. Path to directory for temporary
files generated by file caching.}

\item{cores}{numeric. Maximum number of processing cores. Needs to be set to
1 if file caching is activated.}
}
\value{
R6 sagaGIS object containing a 'gp' attribute containing a nested list
of functions for SAGA-GIS libraries and tools
}
\description{
Dynamically generates functions to all valid SAGA-GIS libraries and tools.
These functions are stored within a sagaGIS R6 object as a named list of 
functions. These functions are accessed via the 'gp' attribute.
}
\examples{
\dontrun{
# initialize Rsagacmd and generate functions for all SAGA-GIS tools
saga = sagaGIS()

# activate file caching for rasters that are too large to fit into memory
saga = sagaGIS(grid_caching=TRUE, grid_cache_threshlod=250, cores=1)

# Example of terrain analysis

# Generate random terrain and load as raster object
dem = saga$gp$grid_calculus$Random_Terrain(RADIUS=100)

# Use Rsagacmd for to calculate the terrain ruggedness index
tri = saga$gp$ta_morphometry$Terrain_Ruggedness_Index_TRI(DEM = dem)
plot(tri)

# Do not load output as an R object
saga$ta_morphometry$geoprocessor$Terrain_Ruggedness_Index_TRI(
    DEM = dem, intern=FALSE)
}
}
