% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsagacmd.R
\name{searchSAGA}
\alias{searchSAGA}
\title{Automatically search for the path to SAGA-GIS.}
\usage{
searchSAGA()
}
\value{
Full path to installed saga_cmd binary
}
\description{
Returns the path to the saga_cmd binary. If multiple versions of SAGA-GIS are
installed on the system, the path to the newest version is returned.
}
