% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{searchTools}
\alias{searchTools}
\title{Search for a SAGA-GIS tool}
\usage{
searchTools(x, pattern)
}
\arguments{
\item{x}{sagaGIS object}

\item{pattern}{character. Pattern of text to search for within the tool name}
}
\value{
list of tools that match the pattern of the search text and their
host library
}
\description{
Search for a SAGA-GIS tool
}
\examples{
\dontrun{
# initialize Rsagacmd
saga = sagaGIS()

# search for a tool
searchTools(saga, 'Terrain')
}
}
