% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tileGeoprocessor}
\alias{tileGeoprocessor}
\title{Split a raster grid into tiles for tile-based processing}
\usage{
tileGeoprocessor(x, grid, nx, ny, overlap = 0)
}
\arguments{
\item{x}{sagaGIS R6 object}

\item{grid}{character, or RasterLayer. GDAL-supported raster, as as a path to
the raster or representing a RasterLayer object}

\item{nx}{numeric. Number of x-pixels per tile}

\item{ny}{numeric. Number of y-pixels per tile}

\item{overlap}{numeric. Number of overlapping pixels}
}
\value{
list. List of RasterLayer objects representing tiled data
}
\description{
Split a raster grid into tiles. The tiles are saved as Rsagacmd
temporary files, and are loaded as a list of R objects for futher
processing. This is a function to make the the SAGA-GIS 
grid_tools / tiling tool more convenient to use.
}
\examples{
library(Rsagacmd)

# check that SAGA-GIS is installed in a recognizable location
if(!is.null(sagaSearch())) {
    saga = sagaGIS()
    
    # generate a random DEM
    dem = saga$gp$grid_calculus$Random_Terrain(RADIUS=15, ITERATIONS=500)
    
    # return tiled version of DEM
    tiles = saga$tileGeoprocessor(grid = dem, nx = 20, ny = 20)
}
}
