% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{update_parameter_file}
\alias{update_parameter_file}
\title{Updates a `parameter` object with file paths to the R data objects.}
\usage{
update_parameter_file(param, temp_path = NULL, backend = NULL)
}
\arguments{
\item{param}{A `parameter` object.}

\item{temp_path}{A character specifying the tempdir to use for storage (optional).}

\item{backend}{A character with the raster backend ('raster' or 'terra).}
}
\value{
A `parameter` object with an updated `file` attribute that refers to
  the on-disk file for saga_cmd to access.
}
\description{
Updates a `parameter` object with file paths to the R data objects.
}
\keyword{internal}
